fix(marquee): animation not resetting
i noticed occasionally one side of the marquee would scroll away and just leave the other side of alternating rows visible and scrolling. it seems like this is the result of the prior animation not resetting after a new page of names is loaded, and since the arabic names are generally shorter, that animation will normally complete first.

trying to reset the animation more reliably by switching to re-inserting the DOM tree for the marquee rows and it seems to work - i haven't seen any blank spaces after a few pages of loading